+++ src/main/kotlin/in/kkkev/jjidea/jj/JujutsuRepository.kt
@@ -138,8 +138,12 @@ data class JujutsuRepositoryImpl(
     override fun createContentRevision(fileAtVersion: FileAtVersion) =
         createContentRevision(fileAtVersion.filePath, fileAtVersion.contentLocator)
 
-    override fun getLogEntry(revision: Revision) = logService.getLog(revision).getOrThrow().singleOrNull()
-        ?: throw VcsException("Multiple log entries found for revision $revision")
+    override fun getLogEntry(revision: Revision) = when (revision) {
+        is ChangeId -> logCache[revision]
+        is CommitId -> logCache[revision]
+        is BookmarkName -> logCache[revision]
+        else -> logService.getLog(revision).getOrThrow().singleOrNull()
+    } ?: throw VcsException("No log entry found for revision $revision")
 
     override fun getLogEntry(contentLocator: ContentLocator) = (contentLocator as? Revision)?.let(this::getLogEntry)